refactor: port all sdk tests to integration tests
With this change all sdk tests no longer use docker to test against
ceramic-one but instead use the hermetic test driver.
