+++ sdk/packages/events/package.json
@@ -45,7 +45,7 @@
   },
   "devDependencies": {
     "@didtools/key-did": "^1.0.0",
-    "dids": "^5.0.2"
+    "dids": "^5.0.3"
   },
   "jest": {
     "extensionsToTreatAsEsm": [".ts"],
