+++ sdk/packages/flight-sql-client/Cargo.toml
@@ -39,3 +39,6 @@ napi-build = "2"
 
 [profile.release]
 lto = true
+
+# Keep this package in its own workspace
+[workspace]
