+++ sdk/packages/test-utils/package.json
@@ -34,7 +34,7 @@
     "@noble/hashes": "^1.5.0",
     "caip": "^1.1.1",
     "did-session": "^4.0.0",
-    "dids": "^5.0.2",
+    "dids": "^5.0.3",
     "viem": "^2.21.14"
   },
   "devDependencies": {
