+++ sdk/packages/test-vectors/package.json
@@ -55,7 +55,7 @@
     "caip": "^1.1.1",
     "cborg": "^4.2.4",
     "did-session": "^4.0.0",
-    "dids": "^5.0.2",
+    "dids": "^5.0.3",
     "key-did-resolver": "^4.0.0",
     "multiformats": "^13.3.0",
     "uint8arrays": "^5.1.0"
