--- sdk/tests/c1-integration/README.md
@@ -1,11 +1,0 @@
-# Ceramic-one integration tests
-
-## Setup
-
-Docker is needed to run Ceramic One for the tests.
-
-## Running tests
-
-```sh
-pnpm test
-```
