+++ tests/suite/package.json
@@ -22,6 +22,15 @@
     "@aws-sdk/client-api-gateway": "^3.713.0",
     "@aws-sdk/client-dynamodb": "^3.713.0",
     "@aws-sdk/client-lambda": "^3.713.0",
+    "@ceramic-sdk/events": "link:../../sdk/packages/events",
+    "@ceramic-sdk/flight-sql-client": "link:../../sdk/packages/flight-sql-client",
+    "@ceramic-sdk/http-client": "link:../../sdk/packages/http-client",
+    "@ceramic-sdk/identifiers": "link:../../sdk/packages/identifiers",
+    "@ceramic-sdk/model-client": "link:../../sdk/packages/model-client",
+    "@ceramic-sdk/model-instance-client": "link:../../sdk/packages/model-instance-client",
+    "@ceramic-sdk/model-instance-protocol": "link:../../sdk/packages/model-instance-protocol",
+    "@ceramic-sdk/model-protocol": "link:../../sdk/packages/model-protocol",
+    "@ceramic-sdk/stream-client": "link:../../sdk/packages/stream-client",
     "@ceramicnetwork/codecs": "^5.4.0",
     "@ceramicnetwork/common": "^6.4.0",
     "@ceramicnetwork/common-test-utils": "^4.4.0",
@@ -32,18 +41,24 @@
     "@ceramicnetwork/streamid": "^5.6.0",
     "@composedb/client": "^0.8.0",
     "@composedb/devtools": "^0.8.0",
+    "@didtools/codecs": "^3.0.0",
+    "@didtools/key-did": "^1.0.0",
     "@didtools/pkh-ethereum": "^0.5.0",
     "@ipld/dag-cbor": "^7.0.3",
     "@ipld/dag-json": "^10.2.3",
+    "@jest/environment": "^29.7.0",
     "@jest/globals": "^29.7.0",
     "@jest/reporters": "^29.7.0",
     "@stablelib/random": "^1.0.2",
+    "@types/cross-spawn": "^6.0.0",
+    "apache-arrow": "18.0.0",
     "axios": "^0.24.0",
     "caip": "^1.1.1",
     "cartonne": "^3.0.1",
     "codeco": "^1.4.3",
     "cross-eventsource": "^1.0.0",
     "cross-fetch": "^4.0.0",
+    "cross-spawn": "^7.0.6",
     "did-session": "^3.1.0",
     "dids": "5.0.3",
     "dotenv": "^16.4.7",
@@ -53,6 +68,7 @@
     "key-did-provider-ed25519": "4.0.2",
     "key-did-resolver": "4.0.0",
     "luxon": "3.2.1",
+    "modern-spawn": "^1.0.0",
     "multiformats": "^13.3.1",
     "multihashes-sync": "^2.0.0",
     "node-jq": "^2.3.5",
