+++ tests/suite/src/__tests__/correctness/fast/model-mid-setType.test.ts
@@ -1,43 +1,43 @@
+import { beforeAll, describe, expect, test } from '@jest/globals'
 import {
   type ClientOptions,
   type FlightSqlClient,
   createFlightSqlClient,
 } from '@ceramic-sdk/flight-sql-client'
 import { CeramicClient } from '@ceramic-sdk/http-client'
-import type { CommitID } from '@ceramic-sdk/identifiers'
-import type { StreamID } from '@ceramic-sdk/identifiers'
 import { ModelClient } from '@ceramic-sdk/model-client'
 import { ModelInstanceClient } from '@ceramic-sdk/model-instance-client'
 import type { ModelDefinition } from '@ceramic-sdk/model-protocol'
-import { getAuthenticatedDID } from '@didtools/key-did'
-import CeramicOneContainer, {
-  waitForEventState,
-  type EnvironmentOptions,
-} from '../src'
+import { waitForEventState } from '../../../utils/rustCeramicHelpers'
+import { urlsToEndpoint } from '../../../utils/common'
+import { randomDID } from '../../../utils/didHelper'
+import { StreamID } from '@ceramic-sdk/identifiers'
 
-const authenticatedDID = await getAuthenticatedDID(new Uint8Array(32))
+const CeramicUrls = String(process.env.CERAMIC_URLS).split(',')
+const CeramicFlightUrls = String(process.env.CERAMIC_FLIGHT_URLS).split(',')
+const CeramicFlightEndpoints = urlsToEndpoint(CeramicFlightUrls);
 
 const testModel: ModelDefinition = {
   version: '2.0',
-  name: 'ListTestModel',
-  description: 'List Test model',
-  accountRelation: { type: 'list' },
-  interface: false,
-  implements: [],
+  name: 'SetTestModel',
+  description: 'Set Test model',
+  accountRelation: {
+    type: 'set',
+    fields: ['test'],
+  },
   schema: {
     type: 'object',
+    $schema: 'https://json-schema.org/draft/2020-12/schema',
     properties: {
-      test: { type: 'string', maxLength: 10 },
+      test: {
+        type: 'string',
+      },
     },
+    required: ['test'],
     additionalProperties: false,
   },
-}
-
-const CONTAINER_OPTS: EnvironmentOptions = {
-  containerName: 'ceramic-test-model-MID-list',
-  apiPort: 5222,
-  flightSqlPort: 5223,
-  testPort: 5223,
+  interface: false,
+  implements: [],
 }
 
 const FLIGHT_OPTIONS: ClientOptions = {
@@ -46,33 +46,35 @@ const FLIGHT_OPTIONS: ClientOptions = {
   password: undefined,
   token: undefined,
   tls: false,
-  host: '127.0.0.1',
-  port: CONTAINER_OPTS.flightSqlPort,
+  host: CeramicFlightEndpoints[0].host,
+  port: CeramicFlightEndpoints[0].port,
 }
 
-const client = new CeramicClient({
-  url: `http://127.0.0.1:${CONTAINER_OPTS.apiPort}`,
-})
 
-const modelInstanceClient = new ModelInstanceClient({
-  ceramic: client,
-  did: authenticatedDID,
-})
-
-const modelClient = new ModelClient({
-  ceramic: client,
-  did: authenticatedDID,
-})
-
-describe('model integration test for list model and MID', () => {
-  let c1Container: CeramicOneContainer
+describe('model integration test for set model and MID', () => {
   let modelStream: StreamID
   let flightClient: FlightSqlClient
+  let client: CeramicClient
+  let modelClient: ModelClient
+  let modelInstanceClient: ModelInstanceClient
 
   beforeAll(async () => {
-    c1Container = await CeramicOneContainer.startContainer(CONTAINER_OPTS)
-    modelStream = await modelClient.createDefinition(testModel)
     flightClient = await createFlightSqlClient(FLIGHT_OPTIONS)
+    client = new CeramicClient({
+      url: CeramicUrls[0]
+    })
+
+    modelInstanceClient = new ModelInstanceClient({
+      ceramic: client,
+      did: await randomDID()
+    })
+
+    modelClient = new ModelClient({
+      ceramic: client,
+      did: await randomDID()
+    })
+
+    modelStream = await modelClient.createDefinition(testModel)
   }, 10000)
 
   test('gets correct model definition', async () => {
@@ -85,7 +87,7 @@ describe('model integration test for list model and MID', () => {
   test('creates instance and obtains correct state', async () => {
     const documentStream = await modelInstanceClient.createInstance({
       model: modelStream,
-      content: { test: 'hello' },
+      content: null,
       shouldIndex: true,
     })
     // Use the flightsql stream behavior to ensure the events states have been process before querying their states.
@@ -94,12 +96,12 @@ describe('model integration test for list model and MID', () => {
     const currentState = await modelInstanceClient.getDocumentState(
       documentStream.baseID,
     )
-    expect(currentState.content).toEqual({ test: 'hello' })
+    expect(currentState.content).toBeNull()
   })
   test('updates document and obtains correct state', async () => {
     const documentStream = await modelInstanceClient.createInstance({
       model: modelStream,
-      content: { test: 'hello' },
+      content: null,
       shouldIndex: true,
     })
     // Use the flightsql stream behavior to ensure the events states have been process before querying their states.
@@ -111,8 +113,5 @@ describe('model integration test for list model and MID', () => {
       shouldIndex: true,
     })
     expect(updatedState.content).toEqual({ test: 'world' })
-  })
-  afterAll(async () => {
-    await c1Container.teardown()
   })
 })
