+++ tests/suite/src/utils/ceramicHelpers.ts
@@ -143,3 +143,4 @@ export async function waitForAnchor(
     msgGenerator,
   )
 }
+
